Add files via upload
This project dives into video , audio and text analysis to enable institutions to make quick decisions and analysis reducing the human intervention. It talks about how the data is extracted using APIs leading to video frame , audio and text extraction to analyze the various features to decide the winners of a competition.